--- demo/simple-web-app/src/utils/promiseTimeout.js
@@ -1,9 +1,0 @@
-const promiseTimeout = (ms, promise) =>
-    Promise.race([
-        promise,
-        new Promise((_, reject) => {
-            setTimeout(() => reject(Error(`Exited by timeout ${ms} ms.`)), ms)
-        }),
-    ])
-
-module.exports = promiseTimeout
