--- demo/simple-web-app/src/utils/sleep.js
@@ -1,2 +1,0 @@
-const sleep = (ms = 0) => new Promise((resolve) => setTimeout(resolve, ms))
-module.exports = sleep
