+++ demo/web_p2p_exchange/pics/Buyer_2.png
