+++ demo/web_p2p_exchange/pics/Seller_2.png
