+++ demo/web_p2p_exchange/src/helpers.js
