+++ demo/web_p2p_exchange/src/ton-contracts/SafeMultisigWallet.tvc
