+++ examples/multisig/contracts.js
@@ -1,4 +1,4 @@
-import { loadContract } from "utils";
+const { loadContract } = require ( "utils");
 
 module.exports = {
     MultisigContract: loadContract("solidity/safemultisig/SafeMultisigWallet"),
